refactor(parser): reorganize tests into focused test modules
- Created src/parser/tests/ directory with 6 focused test modules:
  * basic_tests.rs - basic parsing, assignments, simple commands (13 tests)
  * control_flow_tests.rs - if/elif/else, case, for, while, until, functions (2 tests)
  * pipeline_tests.rs - pipeline parsing (2 tests)
  * operator_tests.rs - logical operators (&&, ||) and negation (!) (6 tests)
  * compound_tests.rs - subshells and command groups (2 tests)
  * redirection_tests.rs - I/O redirection parsing (23 tests)
- Removed ~1,141 lines of test code from src/parser/mod.rs
- Fixed glob import warning by changing pub(crate) use to use
- All 48 parser tests pass, total 669 tests pass
- Parser module reduced from 2,940 to ~1,800 lines
