+++ src/parser/mod.rs
@@ -7,7 +7,7 @@ pub mod ast;
 mod control_flow;
 
 pub use ast::{Ast, Redirection, ShellCommand};
-pub(crate) use control_flow::*;
+use control_flow::*;
 
 use super::lexer::Token;
 
@@ -1797,1144 +1797,4 @@ fn parse_pipeline(tokens: &[Token]) -> Result<Ast, String> {
 
 
 #[cfg(test)]
-mod tests {
-    use super::super::lexer::Token;
-    use super::*;
-
-    #[test]
-    fn test_single_command() {
-        let tokens = vec![Token::Word("ls".to_string())];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["ls".to_string()],
-                redirections: Vec::new(),
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_command_with_args() {
-        let tokens = vec![
-            Token::Word("ls".to_string()),
-            Token::Word("-la".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["ls".to_string(), "-la".to_string()],
-                redirections: Vec::new(),
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_pipeline() {
-        let tokens = vec![
-            Token::Word("ls".to_string()),
-            Token::Pipe,
-            Token::Word("grep".to_string()),
-            Token::Word("txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![
-                ShellCommand {
-                    args: vec!["ls".to_string()],
-                    redirections: Vec::new(),
-                    compound: None,
-                },
-                ShellCommand {
-                    args: vec!["grep".to_string(), "txt".to_string()],
-                    redirections: Vec::new(),
-                    compound: None,
-                }
-            ])
-        );
-    }
-
-    #[test]
-    fn test_input_redirection() {
-        let tokens = vec![
-            Token::Word("cat".to_string()),
-            Token::RedirIn,
-            Token::Word("input.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["cat".to_string()],
-                redirections: vec![Redirection::Input("input.txt".to_string())],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_output_redirection() {
-        let tokens = vec![
-            Token::Word("printf".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RedirOut,
-            Token::Word("output.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["printf".to_string(), "hello".to_string()],
-                compound: None,
-                redirections: vec![Redirection::Output("output.txt".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_append_redirection() {
-        let tokens = vec![
-            Token::Word("printf".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RedirAppend,
-            Token::Word("output.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["printf".to_string(), "hello".to_string()],
-                compound: None,
-                redirections: vec![Redirection::Append("output.txt".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_complex_pipeline_with_redirections() {
-        let tokens = vec![
-            Token::Word("cat".to_string()),
-            Token::RedirIn,
-            Token::Word("input.txt".to_string()),
-            Token::Pipe,
-            Token::Word("grep".to_string()),
-            Token::Word("pattern".to_string()),
-            Token::Pipe,
-            Token::Word("sort".to_string()),
-            Token::RedirOut,
-            Token::Word("output.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![
-                ShellCommand {
-                    args: vec!["cat".to_string()],
-                    compound: None,
-                    redirections: vec![Redirection::Input("input.txt".to_string())],
-                },
-                ShellCommand {
-                    args: vec!["grep".to_string(), "pattern".to_string()],
-                    compound: None,
-                    redirections: Vec::new(),
-                },
-                ShellCommand {
-                    args: vec!["sort".to_string()],
-                    redirections: vec![Redirection::Output("output.txt".to_string())],
-                    compound: None,
-                }
-            ])
-        );
-    }
-
-    #[test]
-    fn test_empty_tokens() {
-        let tokens = vec![];
-        let result = parse(tokens);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "No commands found");
-    }
-
-    #[test]
-    fn test_only_pipe() {
-        let tokens = vec![Token::Pipe];
-        let result = parse(tokens);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "No commands found");
-    }
-
-    #[test]
-    fn test_redirection_without_file() {
-        // Parser doesn't check for missing file, just skips if no token after
-        let tokens = vec![Token::Word("cat".to_string()), Token::RedirIn];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["cat".to_string()],
-                compound: None,
-                redirections: Vec::new(),
-            }])
-        );
-    }
-
-    #[test]
-    fn test_multiple_redirections() {
-        let tokens = vec![
-            Token::Word("cat".to_string()),
-            Token::RedirIn,
-            Token::Word("file1.txt".to_string()),
-            Token::RedirOut,
-            Token::Word("file2.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["cat".to_string()],
-                redirections: vec![
-                    Redirection::Input("file1.txt".to_string()),
-                    Redirection::Output("file2.txt".to_string()),
-                ],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_if() {
-        let tokens = vec![
-            Token::If,
-            Token::Word("true".to_string()),
-            Token::Semicolon,
-            Token::Then,
-            Token::Word("printf".to_string()),
-            Token::Word("yes".to_string()),
-            Token::Semicolon,
-            Token::Fi,
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::If {
-            branches,
-            else_branch,
-        } = result
-        {
-            assert_eq!(branches.len(), 1);
-            let (condition, then_branch) = &branches[0];
-            if let Ast::Pipeline(cmds) = &**condition {
-                assert_eq!(cmds[0].args, vec!["true"]);
-            } else {
-                panic!("condition not pipeline");
-            }
-            if let Ast::Pipeline(cmds) = &**then_branch {
-                assert_eq!(cmds[0].args, vec!["printf", "yes"]);
-            } else {
-                panic!("then_branch not pipeline");
-            }
-            assert!(else_branch.is_none());
-        } else {
-            panic!("not if");
-        }
-    }
-
-    #[test]
-    fn test_parse_if_elif() {
-        let tokens = vec![
-            Token::If,
-            Token::Word("false".to_string()),
-            Token::Semicolon,
-            Token::Then,
-            Token::Word("printf".to_string()),
-            Token::Word("no".to_string()),
-            Token::Semicolon,
-            Token::Elif,
-            Token::Word("true".to_string()),
-            Token::Semicolon,
-            Token::Then,
-            Token::Word("printf".to_string()),
-            Token::Word("yes".to_string()),
-            Token::Semicolon,
-            Token::Fi,
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::If {
-            branches,
-            else_branch,
-        } = result
-        {
-            assert_eq!(branches.len(), 2);
-            // First branch: false -> printf no
-            let (condition1, then1) = &branches[0];
-            if let Ast::Pipeline(cmds) = &**condition1 {
-                assert_eq!(cmds[0].args, vec!["false"]);
-            }
-            if let Ast::Pipeline(cmds) = &**then1 {
-                assert_eq!(cmds[0].args, vec!["printf", "no"]);
-            }
-            // Second branch: true -> printf yes
-            let (condition2, then2) = &branches[1];
-            if let Ast::Pipeline(cmds) = &**condition2 {
-                assert_eq!(cmds[0].args, vec!["true"]);
-            }
-            if let Ast::Pipeline(cmds) = &**then2 {
-                assert_eq!(cmds[0].args, vec!["printf", "yes"]);
-            }
-            assert!(else_branch.is_none());
-        } else {
-            panic!("not if");
-        }
-    }
-
-    #[test]
-    fn test_parse_assignment() {
-        let tokens = vec![Token::Word("MY_VAR=test_value".to_string())];
-        let result = parse(tokens).unwrap();
-        if let Ast::Assignment { var, value } = result {
-            assert_eq!(var, "MY_VAR");
-            assert_eq!(value, "test_value");
-        } else {
-            panic!("not assignment");
-        }
-    }
-
-    #[test]
-    fn test_parse_assignment_quoted() {
-        let tokens = vec![Token::Word("MY_VAR=hello world".to_string())];
-        let result = parse(tokens).unwrap();
-        if let Ast::Assignment { var, value } = result {
-            assert_eq!(var, "MY_VAR");
-            assert_eq!(value, "hello world");
-        } else {
-            panic!("not assignment");
-        }
-    }
-
-    #[test]
-    fn test_parse_assignment_invalid() {
-        // Variable name starting with number should not be parsed as assignment
-        let tokens = vec![Token::Word("123VAR=value".to_string())];
-        let result = parse(tokens).unwrap();
-        if let Ast::Pipeline(cmds) = result {
-            assert_eq!(cmds[0].args, vec!["123VAR=value"]);
-        } else {
-            panic!("should be parsed as pipeline");
-        }
-    }
-
-    #[test]
-    fn test_parse_function_definition() {
-        let tokens = vec![
-            Token::Word("myfunc".to_string()),
-            Token::LeftParen,
-            Token::RightParen,
-            Token::LeftBrace,
-            Token::Word("echo".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RightBrace,
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::FunctionDefinition { name, body } = result {
-            assert_eq!(name, "myfunc");
-            // Body should be a pipeline with echo hello
-            if let Ast::Pipeline(cmds) = *body {
-                assert_eq!(cmds[0].args, vec!["echo", "hello"]);
-            } else {
-                panic!("function body should be a pipeline");
-            }
-        } else {
-            panic!("should be parsed as function definition");
-        }
-    }
-
-    #[test]
-    fn test_parse_function_definition_empty() {
-        let tokens = vec![
-            Token::Word("emptyfunc".to_string()),
-            Token::LeftParen,
-            Token::RightParen,
-            Token::LeftBrace,
-            Token::RightBrace,
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::FunctionDefinition { name, body } = result {
-            assert_eq!(name, "emptyfunc");
-            // Empty body should default to true command
-            if let Ast::Pipeline(cmds) = *body {
-                assert_eq!(cmds[0].args, vec!["true"]);
-            } else {
-                panic!("function body should be a pipeline");
-            }
-        } else {
-            panic!("should be parsed as function definition");
-        }
-    }
-
-    #[test]
-    fn test_parse_function_definition_legacy_format() {
-        // Test backward compatibility with parentheses in the function name
-        let tokens = vec![
-            Token::Word("legacyfunc()".to_string()),
-            Token::LeftBrace,
-            Token::Word("echo".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RightBrace,
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::FunctionDefinition { name, body } = result {
-            assert_eq!(name, "legacyfunc");
-            // Body should be a pipeline with echo hello
-            if let Ast::Pipeline(cmds) = *body {
-                assert_eq!(cmds[0].args, vec!["echo", "hello"]);
-            } else {
-                panic!("function body should be a pipeline");
-            }
-        } else {
-            panic!("should be parsed as function definition");
-        }
-    }
-
-    #[test]
-    fn test_parse_local_assignment() {
-        let tokens = vec![Token::Local, Token::Word("MY_VAR=test_value".to_string())];
-        let result = parse(tokens).unwrap();
-        if let Ast::LocalAssignment { var, value } = result {
-            assert_eq!(var, "MY_VAR");
-            assert_eq!(value, "test_value");
-        } else {
-            panic!("should be parsed as local assignment");
-        }
-    }
-
-    #[test]
-    fn test_parse_local_assignment_separate_tokens() {
-        let tokens = vec![
-            Token::Local,
-            Token::Word("MY_VAR".to_string()),
-            Token::Word("test_value".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::LocalAssignment { var, value } = result {
-            assert_eq!(var, "MY_VAR");
-            assert_eq!(value, "test_value");
-        } else {
-            panic!("should be parsed as local assignment");
-        }
-    }
-
-    #[test]
-    fn test_parse_local_assignment_invalid_var_name() {
-        // Variable name starting with number should not be parsed as local assignment
-        let tokens = vec![Token::Local, Token::Word("123VAR=value".to_string())];
-        let result = parse(tokens);
-        // Should return an error since 123VAR is not a valid variable name
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_here_document_redirection() {
-        let tokens = vec![
-            Token::Word("cat".to_string()),
-            Token::RedirHereDoc("EOF".to_string(), false),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["cat".to_string()],
-                redirections: vec![Redirection::HereDoc("EOF".to_string(), "false".to_string())],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_here_string_redirection() {
-        let tokens = vec![
-            Token::Word("grep".to_string()),
-            Token::RedirHereString("pattern".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["grep".to_string()],
-                compound: None,
-                redirections: vec![Redirection::HereString("pattern".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_mixed_redirections() {
-        let tokens = vec![
-            Token::Word("cat".to_string()),
-            Token::RedirIn,
-            Token::Word("file.txt".to_string()),
-            Token::RedirHereString("fallback".to_string()),
-            Token::RedirOut,
-            Token::Word("output.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["cat".to_string()],
-                compound: None,
-                redirections: vec![
-                    Redirection::Input("file.txt".to_string()),
-                    Redirection::HereString("fallback".to_string()),
-                    Redirection::Output("output.txt".to_string()),
-                ],
-            }])
-        );
-    }
-
-    // ===== File Descriptor Redirection Tests =====
-
-    #[test]
-    fn test_parse_fd_input_redirection() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdIn(3, "input.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                redirections: vec![Redirection::FdInput(3, "input.txt".to_string())],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_output_redirection() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdOut(2, "errors.log".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![Redirection::FdOutput(2, "errors.log".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_append_redirection() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdAppend(2, "errors.log".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![Redirection::FdAppend(2, "errors.log".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_duplicate() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdDup(2, 1),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![Redirection::FdDuplicate(2, 1)],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_close() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdClose(2),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![Redirection::FdClose(2)],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_input_output() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdInOut(3, "file.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![Redirection::FdInputOutput(3, "file.txt".to_string())],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_multiple_fd_redirections() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdOut(2, "err.log".to_string()),
-            Token::RedirectFdIn(3, "input.txt".to_string()),
-            Token::RedirectFdAppend(4, "append.log".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                compound: None,
-                redirections: vec![
-                    Redirection::FdOutput(2, "err.log".to_string()),
-                    Redirection::FdInput(3, "input.txt".to_string()),
-                    Redirection::FdAppend(4, "append.log".to_string()),
-                ],
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_swap_pattern() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdDup(3, 1),
-            Token::RedirectFdDup(1, 2),
-            Token::RedirectFdDup(2, 3),
-            Token::RedirectFdClose(3),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                redirections: vec![
-                    Redirection::FdDuplicate(3, 1),
-                    Redirection::FdDuplicate(1, 2),
-                    Redirection::FdDuplicate(2, 3),
-                    Redirection::FdClose(3),
-                ],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_mixed_basic_and_fd_redirections() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirOut,
-            Token::Word("output.txt".to_string()),
-            Token::RedirectFdDup(2, 1),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                redirections: vec![
-                    Redirection::Output("output.txt".to_string()),
-                    Redirection::FdDuplicate(2, 1),
-                ],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_redirection_ordering() {
-        // Test that redirections are preserved in left-to-right order
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdOut(2, "first.log".to_string()),
-            Token::RedirOut,
-            Token::Word("second.txt".to_string()),
-            Token::RedirectFdDup(2, 1),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["command".to_string()],
-                redirections: vec![
-                    Redirection::FdOutput(2, "first.log".to_string()),
-                    Redirection::Output("second.txt".to_string()),
-                    Redirection::FdDuplicate(2, 1),
-                ],
-                compound: None,
-            }])
-        );
-    }
-
-    #[test]
-    fn test_parse_fd_redirection_with_pipe() {
-        let tokens = vec![
-            Token::Word("command".to_string()),
-            Token::RedirectFdDup(2, 1),
-            Token::Pipe,
-            Token::Word("grep".to_string()),
-            Token::Word("error".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![
-                ShellCommand {
-                    args: vec!["command".to_string()],
-                    redirections: vec![Redirection::FdDuplicate(2, 1)],
-                    compound: None,
-                },
-                ShellCommand {
-                    args: vec!["grep".to_string(), "error".to_string()],
-                    compound: None,
-                    redirections: Vec::new(),
-                }
-            ])
-        );
-    }
-
-    #[test]
-    fn test_parse_all_fd_numbers() {
-        // Test fd 0
-        let tokens = vec![
-            Token::Word("cmd".to_string()),
-            Token::RedirectFdIn(0, "file".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::Pipeline(cmds) = result {
-            assert_eq!(
-                cmds[0].redirections[0],
-                Redirection::FdInput(0, "file".to_string())
-            );
-        } else {
-            panic!("Expected Pipeline");
-        }
-
-        // Test fd 9
-        let tokens = vec![
-            Token::Word("cmd".to_string()),
-            Token::RedirectFdOut(9, "file".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        if let Ast::Pipeline(cmds) = result {
-            assert_eq!(
-                cmds[0].redirections[0],
-                Redirection::FdOutput(9, "file".to_string())
-            );
-        } else {
-            panic!("Expected Pipeline");
-        }
-    }
-
-    // ===== Operator Precedence and Negation Tests =====
-
-    #[test]
-    fn test_negation_with_and_operator() {
-        // Test: ! cmd1 && cmd2
-        // Should parse as: (! cmd1) && cmd2
-        let tokens = vec![
-            Token::Bang,
-            Token::Word("false".to_string()),
-            Token::And,
-            Token::Word("echo".to_string()),
-            Token::Word("success".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be an And node with negation on the left
-        if let Ast::And { left, right } = result {
-            // Left should be a negation
-            if let Ast::Negation { command } = *left {
-                if let Ast::Pipeline(cmds) = *command {
-                    assert_eq!(cmds[0].args, vec!["false"]);
-                } else {
-                    panic!("Expected Pipeline in negation");
-                }
-            } else {
-                panic!("Expected Negation on left side of And");
-            }
-
-            // Right should be echo success
-            if let Ast::Pipeline(cmds) = *right {
-                assert_eq!(cmds[0].args, vec!["echo", "success"]);
-            } else {
-                panic!("Expected Pipeline on right side of And");
-            }
-        } else {
-            panic!("Expected And node, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_negation_with_or_operator() {
-        // Test: ! cmd1 || cmd2
-        // Should parse as: (! cmd1) || cmd2
-        let tokens = vec![
-            Token::Bang,
-            Token::Word("true".to_string()),
-            Token::Or,
-            Token::Word("echo".to_string()),
-            Token::Word("fallback".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be an Or node with negation on the left
-        if let Ast::Or { left, right } = result {
-            // Left should be a negation
-            if let Ast::Negation { command } = *left {
-                if let Ast::Pipeline(cmds) = *command {
-                    assert_eq!(cmds[0].args, vec!["true"]);
-                } else {
-                    panic!("Expected Pipeline in negation");
-                }
-            } else {
-                panic!("Expected Negation on left side of Or");
-            }
-
-            // Right should be echo fallback
-            if let Ast::Pipeline(cmds) = *right {
-                assert_eq!(cmds[0].args, vec!["echo", "fallback"]);
-            } else {
-                panic!("Expected Pipeline on right side of Or");
-            }
-        } else {
-            panic!("Expected Or node, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_negation_and_semicolon_sequence() {
-        // Test: ! cmd1 && cmd2 ; cmd3
-        // Should parse as: Sequence([(! cmd1) && cmd2, cmd3])
-        let tokens = vec![
-            Token::Bang,
-            Token::Word("false".to_string()),
-            Token::And,
-            Token::Word("echo".to_string()),
-            Token::Word("second".to_string()),
-            Token::Semicolon,
-            Token::Word("echo".to_string()),
-            Token::Word("third".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be a Sequence with two commands
-        if let Ast::Sequence(commands) = result {
-            assert_eq!(commands.len(), 2);
-
-            // First command should be (! false) && echo second
-            if let Ast::And { left, right } = &commands[0] {
-                if let Ast::Negation { command } = &**left {
-                    if let Ast::Pipeline(cmds) = &**command {
-                        assert_eq!(cmds[0].args, vec!["false"]);
-                    } else {
-                        panic!("Expected Pipeline in negation");
-                    }
-                } else {
-                    panic!("Expected Negation");
-                }
-
-                if let Ast::Pipeline(cmds) = &**right {
-                    assert_eq!(cmds[0].args, vec!["echo", "second"]);
-                } else {
-                    panic!("Expected Pipeline");
-                }
-            } else {
-                panic!("Expected And node");
-            }
-
-            // Second command should be echo third
-            if let Ast::Pipeline(cmds) = &commands[1] {
-                assert_eq!(cmds[0].args, vec!["echo", "third"]);
-            } else {
-                panic!("Expected Pipeline");
-            }
-        } else {
-            panic!("Expected Sequence, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_nested_logical_operators() {
-        // Test: cmd1 && ! cmd2 || cmd3
-        // Should parse as: (cmd1 && (! cmd2)) || cmd3
-        let tokens = vec![
-            Token::Word("true".to_string()),
-            Token::And,
-            Token::Bang,
-            Token::Word("false".to_string()),
-            Token::Or,
-            Token::Word("echo".to_string()),
-            Token::Word("fallback".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be an Or node
-        if let Ast::Or { left, right } = result {
-            // Left should be: true && ! false
-            if let Ast::And {
-                left: and_left,
-                right: and_right,
-            } = *left
-            {
-                // and_left should be true
-                if let Ast::Pipeline(cmds) = *and_left {
-                    assert_eq!(cmds[0].args, vec!["true"]);
-                } else {
-                    panic!("Expected Pipeline");
-                }
-
-                // and_right should be ! false
-                if let Ast::Negation { command } = *and_right {
-                    if let Ast::Pipeline(cmds) = *command {
-                        assert_eq!(cmds[0].args, vec!["false"]);
-                    } else {
-                        panic!("Expected Pipeline in negation");
-                    }
-                } else {
-                    panic!("Expected Negation");
-                }
-            } else {
-                panic!("Expected And node on left side of Or");
-            }
-
-            // Right should be echo fallback
-            if let Ast::Pipeline(cmds) = *right {
-                assert_eq!(cmds[0].args, vec!["echo", "fallback"]);
-            } else {
-                panic!("Expected Pipeline");
-            }
-        } else {
-            panic!("Expected Or node, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_subshell_with_and_operator_and_sequence() {
-        // Test: (cmd1) && cmd2 ; cmd3
-        // Should parse as: Sequence([(cmd1) && cmd2, cmd3])
-        let tokens = vec![
-            Token::LeftParen,
-            Token::Word("true".to_string()),
-            Token::RightParen,
-            Token::And,
-            Token::Word("echo".to_string()),
-            Token::Word("second".to_string()),
-            Token::Semicolon,
-            Token::Word("echo".to_string()),
-            Token::Word("third".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be a Sequence
-        if let Ast::Sequence(commands) = result {
-            assert_eq!(commands.len(), 2);
-
-            // First should be (true) && echo second
-            if let Ast::And { .. } = &commands[0] {
-                // Structure is correct
-            } else {
-                panic!("Expected And node");
-            }
-
-            // Second should be echo third
-            if let Ast::Pipeline(cmds) = &commands[1] {
-                assert_eq!(cmds[0].args, vec!["echo", "third"]);
-            } else {
-                panic!("Expected Pipeline");
-            }
-        } else {
-            panic!("Expected Sequence, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_command_group_with_or_operator_and_sequence() {
-        // Test: { cmd1; } || cmd2 ; cmd3
-        // Should parse as: Sequence([{ cmd1; } || cmd2, cmd3])
-        let tokens = vec![
-            Token::LeftBrace,
-            Token::Word("false".to_string()),
-            Token::Semicolon,
-            Token::RightBrace,
-            Token::Or,
-            Token::Word("echo".to_string()),
-            Token::Word("second".to_string()),
-            Token::Semicolon,
-            Token::Word("echo".to_string()),
-            Token::Word("third".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be a Sequence
-        if let Ast::Sequence(commands) = result {
-            assert_eq!(commands.len(), 2);
-
-            // First should be { false; } || echo second
-            if let Ast::Or { .. } = &commands[0] {
-                // Structure is correct
-            } else {
-                panic!("Expected Or node");
-            }
-
-            // Second should be echo third
-            if let Ast::Pipeline(cmds) = &commands[1] {
-                assert_eq!(cmds[0].args, vec!["echo", "third"]);
-            } else {
-                panic!("Expected Pipeline");
-            }
-        } else {
-            panic!("Expected Sequence, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_multiple_and_operators_in_sequence() {
-        // Test: cmd1 && cmd2 && cmd3
-        // Should parse as: (cmd1 && cmd2) && cmd3 (left-associative)
-        let tokens = vec![
-            Token::Word("true".to_string()),
-            Token::And,
-            Token::Word("echo".to_string()),
-            Token::Word("second".to_string()),
-            Token::And,
-            Token::Word("echo".to_string()),
-            Token::Word("third".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be an And node with left-associative structure
-        if let Ast::And { left, right } = result {
-            // Left should be: true && echo second
-            if let Ast::And {
-                left: inner_left,
-                right: inner_right,
-            } = *left
-            {
-                if let Ast::Pipeline(cmds) = *inner_left {
-                    assert_eq!(cmds[0].args, vec!["true"]);
-                } else {
-                    panic!("Expected Pipeline");
-                }
-
-                if let Ast::Pipeline(cmds) = *inner_right {
-                    assert_eq!(cmds[0].args, vec!["echo", "second"]);
-                } else {
-                    panic!("Expected Pipeline");
-                }
-            } else {
-                panic!("Expected nested And node on left");
-            }
-
-            // Right should be: echo third
-            if let Ast::Pipeline(cmds) = *right {
-                assert_eq!(cmds[0].args, vec!["echo", "third"]);
-            } else {
-                panic!("Expected Pipeline on right");
-            }
-        } else {
-            panic!("Expected And node, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_negation_in_pipeline() {
-        // Test: ! cmd1 | cmd2
-        // Negation should apply to the entire pipeline
-        let tokens = vec![
-            Token::Bang,
-            Token::Word("grep".to_string()),
-            Token::Word("pattern".to_string()),
-            Token::Pipe,
-            Token::Word("wc".to_string()),
-            Token::Word("-l".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-
-        // Should be a Negation wrapping a Pipeline
-        if let Ast::Negation { command } = result {
-            if let Ast::Pipeline(cmds) = *command {
-                assert_eq!(cmds.len(), 2);
-                assert_eq!(cmds[0].args, vec!["grep", "pattern"]);
-                assert_eq!(cmds[1].args, vec!["wc", "-l"]);
-            } else {
-                panic!("Expected Pipeline in negation");
-            }
-        } else {
-            panic!("Expected Negation, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_redirclobber_without_filename() {
-        // Test that >| without a filename returns an error
-        let tokens = vec![
-            Token::Word("echo".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RedirOutClobber,
-        ];
-        let result = parse(tokens);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "expected filename after >|");
-    }
-
-    #[test]
-    fn test_redirclobber_with_non_word_token() {
-        // Test that >| followed by a non-Word token returns an error
-        let tokens = vec![
-            Token::Word("echo".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RedirOutClobber,
-            Token::Pipe,
-        ];
-        let result = parse(tokens);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "expected filename after >|");
-    }
-
-    #[test]
-    fn test_redirclobber_with_valid_filename() {
-        // Test that >| with a valid filename works correctly
-        let tokens = vec![
-            Token::Word("echo".to_string()),
-            Token::Word("hello".to_string()),
-            Token::RedirOutClobber,
-            Token::Word("output.txt".to_string()),
-        ];
-        let result = parse(tokens).unwrap();
-        assert_eq!(
-            result,
-            Ast::Pipeline(vec![ShellCommand {
-                args: vec!["echo".to_string(), "hello".to_string()],
-                redirections: vec![Redirection::OutputClobber("output.txt".to_string())],
-                compound: None,
-            }])
-        );
-    }
-}
+mod tests;
