+++ src/parser/tests/mod.rs
@@ -1,0 +1,8 @@
+//! Parser test modules
+
+mod basic_tests;
+mod control_flow_tests;
+mod pipeline_tests;
+mod operator_tests;
+mod compound_tests;
+mod redirection_tests;
